synthscholar-comet/requirements.txt
flask==2.3.3
selenium==4.15.0
openai==1.3.0
gtts==2.3.2
python-dotenv==1.0.0
requests==2.31.0
pydub==0.25.1
werkzeug==2.3.7
webdriver-manager==4.0.1
playsound==1.2.2